requirements.txt
langchain
openai
chromadb
faiss-cpu
sentence-transformers
pypdf
streamlit
gradio
requests
python-dotenv